==> scripts/00_ena_template_samples.R <==
# Format ENA sample submission template

# Load libraries
library(tidyverse)
library(janitor)


#### Import and format data ####
# Import environmental data for all lakes
metadata <- read_csv("data/environmental/lakepulse_data_curated_2022-04-28.csv", col_names = TRUE,
                     locale = locale(encoding = "UTF-8"))

# Load sequence table generated by dada2 (ASVs inferred by pooling samples)
load("output/dada2/lp2017-2019watercolumn18s_seqtab_nochim.rda")

# Create list of lakes with sequencing data
lakes <- rownames(seqtab.nochim) %>%
  str_remove_all(., ".*_")
rm(seqtab.nochim)

# Import empty ENA template
# tax_id = 1647806, scientific_name = lake water metagenome
ena <- read_tsv("data/ena/sample_template.tsv", col_names = TRUE, comment = "#") %>%
  remove_empty()

#### Format ENA template submission ####
# Subset metadata for lakes with sequencing data
metadata_lakes <- metadata %>%
  filter(lakepulse_id %in% lakes)

# Select data and rename columns to match template
metadata_lakes <- metadata_lakes %>%
  select(lakepulse_id, sampling_date, lake_name, latitude, longitude, ecozone)

# Combine fields
template <- ena %>%
  bind_cols(metadata_lakes) %>%
  rename(sample_alias = lakepulse_id,
         collection_date = sampling_date,
         `geographic location (latitude)` = latitude,
         `geographic location (longitude)` = longitude,
         `geographic location (region and locality)` = ecozone) %>%
  mutate(common_name = NA,
         sample_description = NA) %>%
  select(sample_alias,
         tax_id,
         scientific_name,
         common_name,
         sample_title,
         sample_description,
         collection_date,
         isolation_source,
         `geographic location (country and/or sea)`,
         `geographic location (latitude)`,
         `geographic location (longitude)`,
         `geographic location (region and locality)`,
         environmental_sample,
         lake_name) %>%
  arrange(sample_alias)

# Add unique prefix to lake IDs
template <- template %>%
  mutate(sample_alias = str_c("LakePulse18S_water_", sample_alias))

# Write template data to file
# template %>%
#   write_tsv("output/ena/lp2017-2019watercolumn18s_ena_template.tsv")
